usn.py: better handle USNs for sources/binaries with epochs
Implement get_best_version() and have read_usn_db() for "allbinaries"
(since historic USNs after December 2018 and before September 2019
incorrectly populated the versions such that the versions always were
set to the source version) and for older USNs without "allbinaries"
where we must guess the binary version due to lack of info available to
the tool. The new algorithm is essentially:

  if allbinaries present:
    get_best_version for binary/version in allbinaries
  else:
    guess the binary version from the URL and and use epoch of the source
    get_best_version for guessed version

get_best_version then consults data/ubuntu-unmatched-bin-versions.json
which is populated with binaries whose version differs from the source
and determines if it should apply/strip an epoch based on looking at the
source_version and binary version in the json file. This is valid since
security updates will not add/modify an epoch and because it is
detectable. Since, binary versions can be anything, we skip binaries
with different versions that we can't accurately guess.