overrides.py: add classic_interfaces_exception and interfaces_attribs_addons
As a special exception, some classic snaps are allowed to specify slots
and plugs. Update sr_lint.py for this and add juju and microk8s to this
list of exceptions.

Additionally, some snaps may provide unknown attributes as a workaround
for coordinating between two snaps, as is the case for microk8s using
'name' as a slot attribute for juju to examine in install hooks.