bin/snap-review: support --state-input and --state-output
Normal flow:

$ snap-review --state-output=out <snap>

Use of --state-output without --state-input will initialize the state
object to a default state that is meaningful to the review-tools. This
is how the store invokes the review-tools when there is no existing
state (eg, new snap, bootstrapping existing snaps). After all the checks
are run, write to the file specified in --state-output (removing the
file if it already exists).

$ snap-review --state-input=in --state-output=out <snap>

Use of --state-output with --state-input will read in the file specified
by --state-input and after all checks are run, write to the file
specified in --state-output (removing the file if it already exists).

Errors related to reading/parsing/validating --state-input before the
normal checks cause snap-review to fail early (using the appropriate
"runtime-errors" json output when --json is specified (ie, just like
other runtime failures).

Errors after the normal checks are run related to validating the
internal state output or writing to --state-output cause snap-review to
add "runtime-errors" to the json output when --json is specified so we
don't lose the checks output.

An error in validating the internal state output prior to writing to
--state-output causes snap-review to discard the internal state output
and output the --state-input (again, adding an appropriate entry to
"runtime-errors"). In this manner, we can avoid saving state to the
store that will later fail to validate.

Finally, the internal state format will support a 'format' version
internally to manage format changes to its internal state as needed.

References:
https://docs.google.com/document/d/1HEfBVBb5pie4383tPZODyPVufAJNrnTRCYxSek6nCno/edit#heading=h.17b7d6xrozfw