sr_declaration.py: use defaults for missing constraints (LP: #1864103)
Historically we've said "if we have no scoped rules, then it is as if
the snap decl wasn't specified for this constraint". However, we really
need to say is for a particular interface and in all the cstr_types (ie,
installation, connection, auto-connection):

 * if there are no scoped rules or unscoped rules in the snap decl, use
   the base decl
 * if we have scoped rules in the snap decl for any cstr_types, use
   those and defaults for the missing cstr_types
 * if we have no scoped rules but have unscoped rules, use unscoped
   rules and defaults for missing cstr_types

In practical terms, this means that an interface in the base decl that
has an installation constraint and slot side connection constraint (ie,
two things that could cause an automated review), only needs only one in
the snap decl to pass automated review.

Considering the above, we can reduce this to simply: "if the snap
declaration for this interface has any cstr_type, use it and use
defaults for any missing cstr_types".

In the process of developing test cases for this, fix the case when we
have scoped snap decl rules ('snap' in whence) for a deny but had no
matching checks and treat the scoped rule as a bool.

References:
https://bugs.launchpad.net/review-tools/+bug/1864103