reviewtools/sr_declaration.py: fix logic error when comparing lists
_check_attrib() correctly documents that when the interface attribute
from the snap.yaml is a list and snap declaration for the attribute is a
string, then all items from the snap.yaml must match the snap decl (a
regex), but there was a logic error that meant only one was needed to
match to pass automated review.

In practice, snapd would not match and so there should be no chance of
regression in the store (since people would've reported the problem).
Add tests to snapd in https://github.com/snapcore/snapd/pull/8226 and
add corresponding test here.