sr_declaration.py: enforce better written snap decls for installation
To promote writing better snap declarations for installation constraints
with interface attributes, enforce:

 a. all attributes specified in the snap declaration are also specified
    in the snap (like snapd)
 b. diverge from snapd and require that all specified attributes in the
    snap are also in the snap declaration

This makes sure that the following (a) triggers a manual review:

  snap has:
    plugs:
      system-files: null

  snap decl has:
    plugs:
      system-files:
        allow-installation:
          plug-attributes:
            write: /foo

While snapd enforces this and fails the installation, this commit
ensures that the snap doesn't pass review and break itself.

It also makes sure that the following (b) triggers manual review:

  snap has:
    plugs:
      system-files:
        read: /foo
        write: /bar

  snap decl has:
    plugs:
      system-files:
        allow-installation:
          plug-attributes:
            read: /foo

snapd correctly does not enforce this since it assumes the snap
declaration author wrote the declaration as intended. As a result, the
above snap successfully installs when that isn't the intent of the snap
declaration author. Since there isn't a convenient way to express
'write should not be specified' in the snap declaration syntax, this
commit diverges from snapd and makes sure if a snap adds another
attribute at some later point, it is flagged for review.