sr_declaration.py: rearrange comments into preamble and code comments
The reasoning behind the checks was sprinkled throughout the code which
made it difficult to reason about the checks and why they are happening.
Create a preamble at the top of the file to make it easier to under what
is happening and why, and then add/adjust code comments to incorporate
and/or expand on things from the preamble.