store.py: make get_ubuntu_release_from_manifest() more robust
get_ubuntu_release_from_manifest() tries to guess the Ubuntu version to
compare against when calculating snap USN notifications. Before, the
algorithm was if snapcraft-os-release-id and
snapcraft-os-release-version-id both exist, use those to calculate the
release. Otherwise if no installed snaps, default to xenial, else if one
installed snap, map it to an Ubuntu release (where an 'installed snap'
is calculated as a snap whose name is in snap_to_release).

While this algorithm has worked for a long time, it was a bit naive
since it didn't take into account non-ubuntu IDs, like "neon". The code
also predates "base snaps" which is another way to determine the Ubuntu
release. The algorithm has been updated to be:

if snapcraft-os-release-id and snapcraft-os-release-version-id both
exist and matches an ubuntu release, return that release

else if "base" is not specified in the manifest, return "xenial"

else if "base" is in snap_to_release, return the corresponding Ubuntu
release

else (we have an unknown base)
  if we have one installed-snap and it is in snap_to_release, then
  return the corresponding release

  else if snapcraft-os-release-version-id is in the manifest, try to see
  if it matches an Ubuntu-version. If so, use the corresponding Ubuntu
  release (note, we really need the manifest to have
  'snapcraft-os-release-id-like', but it doesn't yet)

  if we still don't have an Ubuntu release:
    if we have no installed snaps, return "xenial"

    elif we have > 1 installed snaps, error (we already check '1
    installed snap', above)

When adding more tests, discovered we weren't actually skipping an
improperly formatted entry in 'installed-snaps', so add a 'continue'
to skip it.

Also add 'core20' to snap_to_release since it was missing