common.py: also use readlink() to check for external py3
Optimize _is_external() to short-circuit if not a symlink and comment on
the use of os.path.realpath().

Add use of os.readlink() so we can verify the target of the symlink
against various patterns.

Snapcraft 4.0 updated the python plugin to treat the python in the
snap's runtime as essentially another virtual environment. To achieve
this, snapcraft sets up symlinks to outside of the snap for python, so
add snapcraft_pats to verify against these allowed target paths.