add bin/snap-verify-declaration
Being able to verify a snap declaration without a snap is useful on its
own, so implement verify_snap_declaration() in sr_declaration.py and
bin/snap-verify-declaration as a small wrapper around it. When we
rewrite create-snap-declaration it is expected that it could also
leverage verify_snap_declaration().

Add unit and functional tests for the behavior.