tests/manual-lzo-part2.sh: retry apt ops if they fail
We keep seeing both in this script and in snapd spread tests bugs where apt
calls made sufficiently fast enough inside a focal lxd container fail with
something like:

Type 'bionic-security' is not known on line 50 in source list /etc/apt/sources.list

which is somehow resolved by just running the same command again. My best bet is
that this is a race condition in the cloud-init stuff that is running when the
container starts, so an alternative solution could be to wait for cloud-init
inside the container to finish, and then run these commands, but this was not
thoroughly debugged.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>