Switch to using numeric user/group ids everywhere
Currently review-tools uses the output of unsquashfs -lls when examining
the contents of a snap (for detecting if files have been removed/added or
checking ownership of files etc). This shows the output with user/group
names - however this translation is done based on the hosts /etc/passwd,
not on the base snap - and so this may not be correct. Furthermore, this
may then change the output even if the actual numeric uid/gid does not
change within the snap. This has been observed for the core snap -
https://dashboard.snapcraft.io/snaps/core/revisions/11169/ - where the
actual contents of the snap has not changed but the review-tools believes
it has.

Instead, this commit switches reviewtools to uses unsquashfs -lln which
outputs numeric uid/gids and then switches all internal uses of file
owner/group names to uid/gids and updates the various tests accordingly.

As this is a significant change and will result in a mismatch (and failure
to parse) of the state information stored by the store (as this old state
uses user/group names but we now expect uid/gids) it may be worth bumping
the review-tools version number and removing all the old state information
from the store to avoid any issues.

Signed-off-by: Alex Murray <alex.murray@canonical.com>