sr-common: reject snaps that have duplicated yaml keys
This commit adds detection when yaml keys in maps are duplicated.
In YAML it is not allowed to have duplicated map keys like:
```
map:
 key:
 key:
```
But unfortunately py-yaml and go-yaml.v2 allow this and do not
warn or error. When moving to go-yaml.v3 we noticed this issue.

It is unclear if any real world snaps are affected by this but
we should make sure we at least to not get more.