Makefile: Add python3-ruamel.yaml as a dependency
This is required for to check for duplicate YAML keys. Also change the
check-deb-deps target to use dpkg -s rather than dpkg -l since a package
can still be in the dpkg cache and hence show up via dpkg -l but dpkg -s
error's out if the package is not actually installed.