Fix snap-check-notices handling of snapcraft 5.0 snaps
When analysing a snap built with snapcraft 5.0, snap-check-notices calls
out to snap-updates-available to get a JSON formatted dict of
data. However, when parsing the snap manifest, if a snapcraft version such
as 5.0 is encountered, this is considered an invalid debian package version
and so the output:

WARN: Invalid Snapcraft version: '5.0' not a valid Debian version

is included in the stderr output from snap-updates-available along with any
actual JSON and this causes the JSON parsing in snap-updates-available to
fail as seen in LP: #1950476.

Fixes: LP: #1950476

Signed-off-by: Alex Murray <alex.murray@canonical.com>